Tests/LibWeb/Text/expected/Clipboard/clipboarditem.txt

types: text/plain,text/uri-list
getType('text/plain'): hello
getType('text/uri-list'): https://example.com
